refactor: move metering RPCs into dedicated crate
Extract metering functionality from base-reth-rpc into a new
base-reth-metering crate. This includes:

- MeteringApi trait and MeteringApiImpl
- meter_bundle and meter_block functions
- MeterBlockResponse and MeterBlockTransactions types
- All associated tests

The runner crate now imports from base-reth-metering directly.
Unused dependencies have been removed from base-reth-rpc.
