+++ crates/client/metering/Cargo.toml
@@ -1,0 +1,51 @@
+[package]
+name = "base-reth-metering"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+description = "Metering RPC for Base node"
+
+[lints]
+workspace = true
+
+[dependencies]
+# workspace
+base-bundles.workspace = true
+
+# reth
+reth.workspace = true
+reth-provider.workspace = true
+reth-primitives-traits.workspace = true
+reth-evm.workspace = true
+reth-optimism-evm.workspace = true
+reth-optimism-chainspec.workspace = true
+reth-optimism-primitives.workspace = true
+
+# alloy
+alloy-primitives.workspace = true
+alloy-consensus.workspace = true
+alloy-eips.workspace = true
+
+# rpc
+jsonrpsee.workspace = true
+
+# misc
+tracing.workspace = true
+eyre.workspace = true
+serde.workspace = true
+
+[dev-dependencies]
+base-reth-test-utils.workspace = true
+reth-db = { workspace = true, features = ["test-utils"] }
+reth-db-common.workspace = true
+reth-optimism-node.workspace = true
+reth-testing-utils.workspace = true
+reth-transaction-pool = { workspace = true, features = ["test-utils"] }
+alloy-genesis.workspace = true
+alloy-rpc-client.workspace = true
+op-alloy-consensus.workspace = true
+rand.workspace = true
+tokio.workspace = true
