+++ crates/client/rpc/src/base/mod.rs
@@ -1,5 +1,2 @@
-pub(crate) mod block;
-pub(crate) mod meter;
-pub(crate) mod meter_rpc;
 pub(crate) mod traits;
 pub(crate) mod types;
